ability to install a gem on multiple rubies
Prior to this commit, a user would not be able to use `rvm::define::gem`
to install a gem with the same name (e.g. 'bundler') across multiple
rubies, as the resource would all have the same namevar (bundler) albeit
different parameters (ruby_version, for example).

This commit adds an additional parameter to the defined type: gem_name.
It defaults to `$name` (to be backwards compatible), but otherwise
allows the user to give an arbitrary namevar to the resource and pass
in the gem name as a parameter.